backend/main/requirements.txt
# Web Framework & Server
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# LangChain & AI
langchain-google-genai>=2.0.0
langchain-core>=0.1.0
langchain-mongodb>=0.1.0
langchain-text-splitters>=0.3.0

# YouTube Processing
yt-dlp>=2024.0.0

# Database
pymongo>=4.6.0

# HTTP Client
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
